+++ src/components/skeletons/layouts/distribution-skeleton.tsx
@@ -17,10 +17,10 @@ export function DistributionSkeleton({ className, isMobile = false }: Distributi
         </CardTitle>
         <div className="space-y-1">
           <CardDescription>
-            <span className="inline-block h-4 w-96 animate-pulse rounded-md bg-primary/10" />
+            <Skeleton className="h-4 w-96" />
           </CardDescription>
           <CardDescription>
-            <span className="inline-block h-4 w-72 animate-pulse rounded-md bg-primary/10" />
+            <Skeleton className="h-4 w-72" />
           </CardDescription>
         </div>
       </CardHeader>